+++ app.js
@@ -31,24 +31,12 @@ var http = require('http');
 var logger = require('morgan');
 var path = require('path');
 
-global.root = path.resolve(__dirname);
 
-// Rutas
-var index = require(root + '/routes/index');
-var infoInstitucional = require(root + '/routes/infoInstitucional');
-var administracion = require(root + '/routes/administracion');
-var docencia = require(root + '/routes/docencia');
-var gestionInvestigacion = require(root + '/routes/gestionInvestigacion');
-var normativaLegal = require(root + '/routes/normativaLegal');
-var solicitudInfo = require(root + '/routes/solicitudInfo');
-var mapaWeb = require(root + '/routes/mapaWeb');
-var buscador = require(root + '/routes/buscador');
-var calendario = require(root + '/routes/calendario');
-var calendarioSolo = require(root + '/routes/calendarioSolo');
-var prueba = require(root + '/routes/prueba');
+var routes=require('./app/routes.js');
+
 
 // Librerías
-var cargar = require(root + '/lib/cargar');
+var cargar = require('./lib/cargar');
 // Prueba de escritura de JSON recuperado de API
 //var escribir = require(__dirname+'/lib/putJSON');
 //escribir();
@@ -56,77 +44,25 @@ var cargar = require(root + '/lib/cargar');
 // Crea aplicación web con Express
 var app = express();
 
+
 // Carga y exporta el archivo de configuración de la aplicación
-config = cargar(root + '/config/config.json');
+config = cargar('./config/config.json');
 module.exports.config = config;
 
 // Archivos de configuración de cada unas de las páginas
-module.exports.personal = cargar(root + '/config/personal.json');
-module.exports.infoEconomica = cargar(root + '/config/infoEconomica.json');
-module.exports.ofertaDemanda = cargar(root + '/config/ofertaDemanda.json');
-module.exports.claustro = cargar(root + '/config/claustro.json');
-module.exports.estudiantes = cargar(root + '/config/estudiantes.json');
-module.exports.gobierno = cargar(root + '/config/gobierno.json');
-module.exports.rendimiento = cargar(root + '/config/rendimiento.json');
-module.exports.normativaLegal = cargar(root + '/config/normativaLegal.json');
-module.exports.prueba = cargar(root + '/config/prueba.json');
+module.exports.personal = cargar('./config/personal.json');
+module.exports.infoEconomica = cargar('./config/infoEconomica.json');
+module.exports.ofertaDemanda = cargar('./config/ofertaDemanda.json');
+module.exports.claustro = cargar('./config/claustro.json');
+module.exports.estudiantes = cargar('./config/estudiantes.json');
+module.exports.gobierno = cargar('./config/gobierno.json');
+module.exports.rendimiento = cargar('./config/rendimiento.json');
+module.exports.normativaLegal = cargar('./config/normativaLegal.json');
+module.exports.prueba = cargar('./config/prueba.json');
 
-// Inicio
-app.get('/', index.index);
-app.get('/index.html', index.index);
-// Información institucional
-app.get('/infoInstitucional.html', infoInstitucional.infoInstitucional);
-// Administracion
-app.get('/personal.html', administracion.personal);
-app.get('/infoEconomica.html', administracion.infoEconomica);
-app.get('/perfilContratante.html', administracion.perfil);
-// Docencia
-app.get('/ofertaDemanda.html', docencia.ofertaDemanda);
-app.get('/claustro.html', docencia.claustro);
-app.get('/estudiantes.html', docencia.estudiantes);
-// Gestion e Investigación
-app.get('/gobierno.html', gestionInvestigacion.gobierno);
-app.get('/rendimiento.html', gestionInvestigacion.rendimiento);
-// Normativa Legal
-app.get('/normativaLegal.html', normativaLegal.normativaLegal);
-// Solicitudes de Información
-app.get('/solicitudInfo.html', solicitudInfo.solicitudInfo);
-// Mapa del sitio
-app.get('/mapaWeb.html', mapaWeb.mapaWeb);
-// Buscador
-app.get('/buscador.html', buscador.buscador);
-// Calendario
-app.get('/calendario.html', calendario.index);
-app.get('/calendarioSolo.html', calendarioSolo.index);
-// Prueba para nueva visualización de tablas
-app.get('/prueba.html', prueba.personal);
-
-// Archivos de datos
-app.get('/archivos/personal', function(req, res) {
-  res.send(cargar(root + '/config/personal.json'));
-});
-app.get('/archivos/infoEconomica', function(req, res) {
-  res.send(cargar(root + '/config/infoEconomica.json'));
-});
-app.get('/archivos/ofertaDemanda', function(req, res) {
-  res.send(cargar(root + '/config/ofertaDemanda.json'));
-});
-app.get('/archivos/claustro', function(req, res) {
-  res.send(cargar(root + '/config/claustro.json'));
-});
-app.get('/archivos/estudiantes', function(req, res) {
-  res.send(cargar(root + '/config/estudiantes.json'));
-});
-app.get('/archivos/gobierno', function(req, res) {
-  res.send(cargar(root + '/config/gobierno.json'));
-});
-app.get('/archivos/rendimiento', function(req, res) {
-  res.send(cargar(root + '/config/rendimiento.json'));
-});
-app.get('/archivos/normativaLegal', function(req, res) {
-  res.send(cargar(root + '/config/normativaLegal.json'));
-});
-
+//will set all roots
+//TODO:maybe 
+routes(app);
 
 // Variables de entorno (puerto de escucha y dirección IP)
 app.set('port', process.env.PORT);
@@ -137,7 +73,7 @@ app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'jade');
 
 // Favicon
-app.use(favicon(root + '/public/favicon/favicon.ico'));
+app.use(favicon('./public/favicon/favicon.ico'));
 // Logger de solicitudes HTTP
 app.use(logger('dev'));
 // Parseadores
